pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ieee2030-5-gateway"
version = "0.1.0"
description = "IEEE 2030.5 DER gateway — Python integration layer over the EPRI client"
requires-python = ">=3.11"

dependencies = [
    "pymodbus>=3.6",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
gateway = "gateway.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]
